requirements.txt
txt
aiogram==3.18.0
python-dotenv==1.0.0
aiohttp==3.9.0
colorama==0.4.6
Flask==2.3.0
gunicorn==20.1.0
